Use a version string matching gitlab repo but not drupal.org
This is re #5 and is probably not as good an approach as using the
drutopia/drutopia namespace (as acquia/lightning does and as Nedjo
recommends) but this puts it back to how it was working before for us.

And Drupal.org just needs to fix https://www.drupal.org/node/2763017
because this is ridiculous.